add SlashSegment, not finished

+++ src/chapter5/c6/ISegment.java
@@ -1,14 +1,14 @@
 package chapter5.c6;
 
 public interface ISegment {
-    public int getStart();
-    public int getEnd();
-    public void setStart(int s);
-    public void setEnd(int e);
-    public ISegment subSeg(int s, int e);
+    public double getStart();
+    public double getEnd();
+    public void setStart(double s);
+    public void setEnd(double e);
+    public ISegment subSeg(double s, double e);
     public interface ISegmentTool{
         public Bean merge(ISegment sA, ISegment sB);
-        public ISegment createFlat(int start,int end, int value);
+        public ISegment createFlat(double start,double end, double value);
         public ISegment concatenate(ISegment sA,ISegment sB);
     }
     public class Bean{
